Fix failing MediaBrowserTest in API >= 28
The previously submitted CL "149c421: Fix throwing Exception
in getCurrentBrowserInfo/getBrowserRootHints" had a bug, which resulted
the MediaBrowserTest started to fail in API >= 28.

The bug was calling MediaBrowserServiceImplApi28#getBrowserRootHints()
in API 28, before the variable mConnection is set.

Since the bug was not caught due to the low readability of the code,
this CL not only fixes the bug, but also does some refactoring.

Bug: 120638640
Test: ./media/version-compat-tests/runtest.sh 4 (in API 28)
Change-Id: I10c887f70028103b6a122130acf318419b419336