Update patch set 5
Patch Set 5:

/usr/local/google/buildbot/src/android/androidx-master-dev/frameworks/support/testutils/src/main/java/androidx/testutils/CountOnDraw.java:42: Error: Access to private field mView of class CountOnDraw requires synthetic accessor [SyntheticAccessor]
mView.getViewTreeObserver().removeOnDrawListener(CountOnDraw.this);
~~~~~
/usr/local/google/buildbot/src/android/androidx-master-dev/frameworks/support/testutils/src/main/java/androidx/testutils/CountOnDraw.java:43: Error: Access to private field mLatch of class CountOnDraw requires synthetic accessor [SyntheticAccessor]
mLatch.countDown();
~~~~~~

Explanation for issues of type "SyntheticAccessor":
A private inner class which is accessed from the outer class will force the
compiler to insert a synthetic accessor; this means that you are causing
extra overhead. This is not important in small projects, but is important
for large apps running up against the 64K method handle limit, and
especially for libraries where you want to make sure your library is as
small as possible for the cases where your library is used in an app
running up against the 64K limit.

2 errors, 0 warnings

You have a lint warning. However, given that this is a test library that we don't ship, you can probably just disable SyntheticAccessor lint check in build.gradle.

Patch-set: 5
Reviewer: Gerrit User 1086749 <1086749@85c56323-6fa9-3386-8a01-6480fb634889>