Fix mock ViewModel and clear()
Previously, developers could mock a
ViewModel and call clear() on the
containing ViewModelStore, which would
result in a call to onCleared() on all
ViewModels.

With the introduction of the final
package private clear() method on ViewModels
that contains a reference to the mBagOfTags
member variable which is null for mock objects,
this behavior broke.

By checking for a null mBagOfTags, we ensure
that the existing behavior of the stable
ViewModels is retained.

Test: new ViewModelTest test
BUG: 122273087
Change-Id: I458da0441cb98c0a4374c0d3e452e1ee1429e158