Update patch set 1
Patch Set 1: Code-Review-1

It is quite sad and not scalable for a future. If we positively allow mocking, then we should always add such artificial checks on every property, even if it is against invariants that we have in code.

Patch-set: 1
Reviewer: Gerrit User 1098853 <1098853@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1