Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> It is quite sad and not scalable for a future. If we positively allow mocking, then we should always add such artificial checks on every property, even if it is against invariants that we have in code.

Mocking is already in use with the stable version of ViewModels - whether you explicitly supported it or not, mocking was supported and no longer is.

Is the plan to move ViewModel to 3.0 to reflect the breaking change? I don't think a 2.1, particularly of something transitively included by other libraries, should force developers to rewrite all of their tests.

Patch-set: 1
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>