Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > It is quite sad and not scalable for a future. If we positively allow mocking, then we should always add such artificial checks on every property, even if it is against invariants that we have in code.
> 
> Mocking is already in use with the stable version of ViewModels - whether you explicitly supported it or not, mocking was supported and no longer is.
> 
> Is the plan to move ViewModel to 3.0 to reflect the breaking change? I don't think a 2.1, particularly of something transitively included by other libraries, should force developers to rewrite all of their tests.

it is a slippery slope though as this was never an API to begin with. As long as we've not promoted this practice (Idk if we did), it is OK to break it if we want to.

That being said, in this particular case, developer seems to be trying to test a fragment by mocking a VM which should work. They are not really calling clear manually as far as I understand.
So I think what we need to make work is that scenario (fragment testing lib working fine w/ mocked view models).

Patch-set: 1
Reviewer: Gerrit User 1060337 <1060337@85c56323-6fa9-3386-8a01-6480fb634889>