Introduce by navArgs() for Fragments
Since getArguments() isn't valid in the
constructor, developers need to use a
lazy operation to create their Args instances.

By adding a common interface, NavArgs, for all
Args classes, we can introduce a property
delegate that takes care of the lazy functionality
and uses reified typing to make for a cleaner API.

As part of this change, the static fromBundle() methods
have been replaced with a constructor that takes a
Bundle in order to simplify ProGuard configuration.

Test: new FragmentTest test, existing safe args tests pass
BUG: 119517986
Change-Id: I02580eba666b4ff0c95ff1c3bff77fe8033e0ba8