Enable media-widget tests for Android K
Fix handling of seeks before preparation in the ExoPlayer wrapper.

Fix calling getCurrentPosition in MediaPlayer.java to ensure that an
exception from getting the position while idle is converted to -1.

Also wrap internal player errors in a RuntimeException so that the outer
function's stack trace is logged in addition to the internal one.

Bug: 80232248
Bug: 122505387
Test: Tests pass locally
Change-Id: I0215e89f9d42dc4cd4d59622409603eada0fcd7d