Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (7 comments)
> 
> resolved comments.
> 
> > Patch Set 1:
> > 
> > Also: is forking really necessary? Is it unacceptable to map unique code to Handler? If that works, is there some way we can implement scheme matching on top of the original UriMatcher?
> > 
> > If there's a reasonably performant and relatively straightforward non-forked solution, I would prefer that to maintaining more code. It just feels wrong for an Android project to fork Android...
> 
> I gave it a quick try when I was first porting the code from github but It failed. I totally agree with you. I'll give it another try.

So the fork also adds the support for the wild card "**" to support nested paths which is not supported by the current android.content.UriMatcher. This is needed for matching nested resources and assets paths if I get this correct.

Here is my trial to use android.content.UriMatcher aosp/866990

Patch-set: 2
Reviewer: Gerrit User 1402701 <1402701@85c56323-6fa9-3386-8a01-6480fb634889>