Fix expected state change sequence + improvements
When determining if we expect the state change event sequence to have a
SETTLING state after DRAGGING, we must base this decision solely on the
offset in pixels of the last scroll event after the DRAGGING event.
The settling phase is skipped if and only if the drag ends exactly on a
page, no matter which page.

2) Go to start page without unnecessary delay
When doing a programmatic scroll without smoothscroll, there are no
state change events, so we shouldn't wait for IDLE. Use the existing
setCurrentItemSync method instead.

3) Add more info to error msg on failed assert

Print the entire event list when DragWhileSmoothScrollTest.test fails,
so we can figure out what's going on.

Bug: 73968249
Test: ./gradlew viewpager2:connectedMaxDepVersionsDebugAndroidTest
Change-Id: I110fa3b3d0a8fd135ee0fe31d6b6154593576dc6