Update patch set 1
Patch Set 1:

I think I remember something. Wasn't it related to AMS and not Truth? Because they link Jetifier's jar to AS binary and there was a clash with ASM versions. I think truth shouldn't be thr issue as that it not published outside of support library project.

Patch-set: 1
Reviewer: Gerrit User 1238792 <1238792@85c56323-6fa9-3386-8a01-6480fb634889>