Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think I remember something. Wasn't it related to AMS and not Truth? Because they link Jetifier's jar to AS binary and there was a clash with ASM versions. I think truth shouldn't be thr issue as that it not published outside of support library project.

Does this mean I shouldn't update it for jetifier as well?  They had separately specified Truth 0.34 without using the constant, but it was the same version and there wasn't a comment or anything explaining why it was done that way, so I assumed it was just an oversight.

Patch-set: 1
Reviewer: Gerrit User 1014468 <1014468@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0