Coroutine scope for Lifecycle
This CL adds a `scope` property to `LifecycleOwner` and
`Lifecycle` classes. It is a CoroutineSCope that gets
cancelled when the Lifecycle is destroyed.

I've also added a `whenStarted` suspend function which
receives a block and runs it when the lifecycle is at
least STARTED state. It suspends until that condition is
met AND cancels if the lifecycle is destroyed while
waiting.

Bug: 122740811
Test: gw :life:integ:l-t-k:cC :life:integ:l-t-k:test

Change-Id: Id565518d9191f21335c793e77d3509b3f3779bbb