Update patch set 8
Patch Set 8:

> Patch Set 7:
> 
> (1 comment)

Hey Insun, the issue is that the old AGP had some issues detecting lint errors of dependencies on local projects. So for example if you depend on a restricted API of "androidx.core:core" then lint would detect that, but if you depend on ":core" then lint would miss that because it was not working properly for projects. The new AGP however works properly on projects (in theory...) so it's detecting existing lint issues more efficiently. This means we have to suppress them when we are updating. Although for now we have decided to wait on this CL and see if we can change the library group restrictions to androidx restrictions (meaning all AndroidX libraries can use that API instead of only the library that contains it.) This way we will not have to suppress the lint errors and we'll probably get rid of most if not all of the suppressions :)

Patch-set: 8
Reviewer: Gerrit User 1354330 <1354330@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0