Update patch set 2
Patch Set 2:

Right. So the plan for this CL is to have three classes up for review:

1. The new ResourceManager (still in the same package as ACDM) which is the 80% of ACDM without that hooks block.
2. The hooks interface
3. ACDM with the implementation of the interface (the green block on the right) + piping all the method implementations to the new ResourceManager.

Once that is in, we can then do a mass move of all other resource-related classes to a new androidx.resources module, with only 3) from above staying in appcompat.

Patch-set: 2
Reviewer: Gerrit User 1061051 <1061051@85c56323-6fa9-3386-8a01-6480fb634889>