Update patch set 2
Patch Set 2:

When I try to access a non-existing file and when I open networks requests in developer tools in chrome, I find that the return status is 200!

At first I thought it's because the PathHandler in WebViewAssetLoader class was returning 200/OK for all requests but after fixing this in the code it's still 200/OK.

I'll try to look into this tomorrow .. but do you have any thoughts?

Patch-set: 2
Reviewer: Gerrit User 1402701 <1402701@85c56323-6fa9-3386-8a01-6480fb634889>