Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > actually should we just go and update the lifecycle version since it is already alpha as well ?
> 
> See my initial comment on the CL:
> "I'm not a particular fan of releasing a new version of all of the many Lifecycle artifacts for no reason, nor am I particularly keen on splitting ViewModels version.
> 
> This seemed to better of the two options since ViewModel is already treated separately from the other Lifecycle artifacts externally."
> 
> Do we want to rev everything? Or split ViewModel?

whoops missed that. I would be OK to rev it separately if lifecycle was not already alpha. but given that it is alpha, i see less harm in doing that. 

Also, idk how this impacts ViewModelSavedState so that might be a reason to rev view models separately.

Ultimately, I would like to keep same group artifacts in the same version except for new artifacts.

Patch-set: 1
Reviewer: Gerrit User 1060337 <1060337@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0