Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)
> 
> > Patch Set 2:
> > 
> > When I try to access a non-existing file and when I open networks requests in developer tools in chrome, I find that the return status is 200!
> > 
> > At first I thought it's because the PathHandler in WebViewAssetLoader class was returning 200/OK for all requests but after fixing this in the code it's still 200/OK.
> > 
> > I'll try to look into this tomorrow .. but do you have any thoughts?
So I found the following (from playing around and just testing the behaviour, haven't looked into WebResourceResponse class source code)

WebResourceResponse seems to handle mimeType, response status and reasonphrase ... etc, so this does not need to be done in WebViewAssetLoader I guess.

WebResourceResponse always overrides the given response status by checking the InputStream:
if null: sets it to 404 not found.
if not null: sets it to 200 OK.

Since LazyInputStream is used in WebViewAssetLoader so an InputStream object is always provided. So WebResourceResponse always sets the response status to 200 regardless of the encapsulated InputStream in LazyInputStream.

> 
> Does the code in WebView have the same behaviour?

In WebView when an non-exist file is accessed (for example "file:///android_asset/nonexist_file.ext") it returns a null WebResourceResponse which is shown as "failed net::ERR_NOT_FOUND" in the status field in the developers console. So if we want to have a similar code we should return null not 404 and we should drop using LazyInputStream.

Patch-set: 4
Reviewer: Gerrit User 1402701 <1402701@85c56323-6fa9-3386-8a01-6480fb634889>