Update patch set 8
Patch Set 8:

(1 comment)

It might not actually be helping to have this split into this many CLs at this point - the CLs all touch a nonoverlapping set of files so it doesn't really make it much clearer, and it makes it hard to see how the parts relate to each other.

It's generally most useful to have a series of CLs when they are iterating on the same code (adding more functionality/etc to the same thing) - it's not as useful if the "earlier" CLs are introducing something that is only used by its own unit tests and don't get changed at all in later CLs.

This is definitely a matter of opinion though ;)

Patch-set: 8
Reviewer: Gerrit User 1021004 <1021004@85c56323-6fa9-3386-8a01-6480fb634889>