Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> Had offline conversations with Paul and Hazem, but I'll post my take-aways:
> 
>  * Chromium has (always?) supported .svgz for network resources

I don't think they will be automatically decompressed unless they are served with "Content-Encoding: gzip". There isn't a separate MIME type for svgz. Web servers probably do the right thing in most cases, though.

>  * Paul implemented .svgz support for file:///android_asset/ URLs in WebView (http://crrev/c/1134429)

Yeah, we implemented this because there's no such thing as a Content-Encoding for file:, since there are no headers at all. The generic code for handling file:// in Chromium already has the same behaviour, it's where we copied this logic from, and the justification for supporting it in file:///android_asset.

>  * I'm not sure if shouldInterceptRequest supports .svgz (Paul's change doesn't affect this).

When it's a network request from Chromium's POV (which includes shouldInterceptRequest, since that returns headers/etc), I believe the expectation is that this will be served with "Content-Encoding: gzip", which will cause the decompression to happen in the network stack automatically, without caring what the file extension actually is. The special case handling only applies to file://.

>    * If it already supports this, then the relevant code here is redundant.

Well, we could try inserting the Content-Encoding header instead I guess, but we'd still have to special case the file extension because the MIME type guessing logic will return the same answer for svgz and svg.

>    * If not, should we consider adding support in shouldInterceptRequest (for the sake of non-AssetLoader use cases)?

No; network requests should rely as little as possible on guessing filetypes and rely on the HTTP headers being correct.

> Then there's the broader question: do we actually want to support .svgz for the AssetLoader API? I don't see obvious problems with supporting it, but the format itself feels kind of unnecessary.

It's fairly common on the web because .svg is an inefficient plaintext format; while many web server setups will automatically compress the assets they're serving (and serve them with the appropriate Content-Encoding header to have the client decompress it), it seems like a lot of people still gzip their svg files in advance, and a bunch of image editors seem to have native support for saving as svgz. We added it because Android developers explicitly asked for it, and because it already worked with actual on-disk files due to the existing special case in chromium for that.

Patch-set: 10