Start using applyOverrideConfiguration WIP
Allows us to fully work with WebViews! The problem
being that we have a VERY small window to call the
method before we have to fall back to the current
behavior.

The last place it can be called is attachBaseContext()
which is before onCreate(). That causes a problem since
we use the savedInstanceState to keep track of the current
local night mode, thus we don't actually know the
correct night mode to apply in attachBaseContext().

This means we have had to move to a static store for
any in-flight local night modes.

Test: included
BUG: 62477814

Change-Id: I0ec61c9e6109f86d26606992e5bcb58fb5bff88f