Fix ChangeImageTransform wrong matrix applied on API < 21 bug
There was a logic to store the current scaleType and matrix before the animation and restore it after the animation for pre-lollipop. It works fine until we start another concurrent transition while the previous one is still in progress:
1) User scheduled the new transition and changed the scaleType to the value he wants
2) Old animation canceled and end listener is called.
3) In this listener we apply our stored scaleType and overrides the value user provided.
4) We start new animation and now store the wrong value - our restored, but not he final value provided by user
5) After animation we restore the wrong scaleType.
To prevent this for happening we should not save and restore but reimplement the ImageView.animateTransform() method logic for API's below 21. We have to use reflection(only before 21) to change the value of ImageView.mDrawMatrix to apply a transient drawing only matrix for a view, it will only be applied for the next drawing and not clear the user provided matrix.

Bug: 123226255
Test: added a test for this case
Change-Id: Id9870adc59ba815bfb37cedc42426cb6921d11df