Media2-Widget: disable subtitle button in MCV when no subtitles
Even the subtitle button is diabled when there is no subtitle tracks, it
is enabled when MCV.setEnabled(true) is called.

So it looks diabled but is clickable and
when it is clicked it tries to show empty PopupWindow.
After that pressing back button stops activity for 5~10 seconds.

Avoid that by checking exsitence of subtitle tracks prior to showing popup
window.

Bug: 123271636
Test: ran support demos manually and do reproduction steps in
b/123271636

Change-Id: I66b80cd66cd8f7d717022f9dcacd41b69844aca5