Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

Yuichi, apologies for taking a while to reply to this CL.

This looks good, I think deprecating these conflict resolutions is currently right given Android's current SQLite implementation in the framework being broken. However, in the future this might be fixed or the user might be using a more up-to-date binding (say a compiled version from sqlite.org and packaged with the app) so I think we shouldn't remove them and in fact we might be able to bring them back if we completely un-bundle SQL.

I filled a follow-up issue b/123313071 so that somehow, sometime in the future we revisit this deprecation since overall we want to have full parity with SQLite's API

+1 now for a small documentation comment, otherwise this is +2 from me.

Thanks!

Patch-set: 1
Reviewer: Gerrit User 1258207 <1258207@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1