Updates annotations for tests previously marked small/medium
This cl does not cover tests previously annotated as @LargeTest that can
be bumped down, nor does it touch tests that do not currently have any
annotations. This cl also only modifies annotations on the class as a
whole, not individual methods. There's a lot of variance in test timing,
so some of these tests may still move around a lot, but this should as
an initial pass cut down presubmit times a lot.

SmallTest = < 0.2 seconds
MediumTest = < 1 second
LargeTest = > 1 second

Bug: b/123402142
Test: ./gradlew connectedCheck --daemon -Pandroid.testInstrumentationRunnerArguments.size=small -Pandroid.testInstrumentationRunnerArguments.timeout_msec=200 --continue
Change-Id: Id9eae3cf720329c0788e27a062d543cc6a2f9852