Update patch set 4
Patch Set 4:

How is this positioned against -night as an automatic side-effect of Battery Saver in SDK 28+? Is this intended to backport the feature to SDK <28? If so, I'm concerned that this is adding unnecessary complexity and disparity between apps.

Is there a b/ where we can continue discussion?

Patch-set: 4
Reviewer: Gerrit User 1044446 <1044446@85c56323-6fa9-3386-8a01-6480fb634889>