Safe Args getApplicationIdTextResource workaround.
For some unknown cases the new Android Gradle Plugin method
getApplicationIdTextResource() returns a null TextResource which causes
the Safe Args plugin to fail in projects using versions of the AGP that
are above 3.2. In order to mitigate the issue this CL will still attempt
to use the old method (getApplicationId) if the TextResource returned is
null.

Manual tests have shown that for the cases where the TextResource
is most needed it does not return null (feature modules) and for those
where the old API is good enough (library or app modules) then it is
safe to call getApplicationId() as it won't throw.

This CL also bumps the Safe Args's AGP dependency to 3.3.0 and KGP to
1.3.20.

Bug: 121304903
Test: Manual tests with AGP 3.3.0 and 3.4.0-beta02 in an app, library
      and feature modules. Also updated PluginTest

Change-Id: Ie33d886092744fa2ce638e47f104ae0fd9cab234