Update patch set 2
Patch Set 2:

Is this primarily for apps that want to use butterknife with reflection? If so, the JSR 305 annotations already use runtime retention so you could use those.

But if this is a request for more general analysis where you need to not just look at your own code but the nullness contracts of libraries (such as the support library) then obviously the support library annotations would need to have runtime retention as well, though it would take a few release cycles until all the libraries you need are compiled with the latest annotation library.

Here's another idea: What if there was a reflection/runtime version of the annotations library, and then you could depend on that version instead inf your debug configuration?

Patch-set: 2
Reviewer: Gerrit User 1004689 <1004689@85c56323-6fa9-3386-8a01-6480fb634889>