Fix night mode recreating Activities handling config changes
Caused by the fact that we call applyDayNight() from
onConfigurationChanged(), which will then call through
to recreate(). We can safely ignore recreating from
onConfigurationChanged(), which is what this CL does.

The easiest way to test this is for Activities handling
configChanges="orientation". Currently we'll recreate
those Activities which is obviously incorrect.

Test: TBD
BUG: 112901684
Change-Id: Ia5edc5cd0e75d7c5749df84f667eee04a9ae4ec6