Update patch set 1
Patch Set 1:

Adding a default implementation that throws an exception isn't actually solving the problem; you have just shifted it elsewhere while technically preserving binary compatibility. In practice, this is worse as it appears to be source- and binary-compatible but is not.

This is not a fix.

Patch-set: 1
Reviewer: Gerrit User 1044446 <1044446@85c56323-6fa9-3386-8a01-6480fb634889>