Set BrowseSupportFragment as primary navigation fragment in fragment manager.
This change allows the child fragment manager to receive all of the back stack change events allowing the fragment to manage the header fragment, main fragment, and transitions between them. This follows the same pattern from NavHostFragment. In addition, BrowseSupportFragment is compatible with FragmentNavigator since both classes are not "fighting" over the same parent fragment manager.

This change should also be copied/pasted into BrowseFragment but since the fragment is deprecated and FragmentTransaction#setPrimaryNavigationFragment is added in API 26, it is quite complex to update the behavior. It is okay for the update to BrowseFragment to happen in it's own CL.

Test: BrowseSupportFragmentTest, Demo App(support-leanback-demos), manually included into the github smaple
Change-Id: Ia7addee2fda0abfd25c013ae35bcbcdccfdf0699