Update patch set 2
Patch Set 2:

I'm not sure what the purpose of this file is, is it to reflect the current status of migration on g3? Because if so then it contains libraries that have not yet migrated (e.g fragment.) I was under the impression that jetifying using this file should only migrate from android support to androidx the libraries that have successfully migrated in g3.

Patch-set: 2
CC: Gerrit User 1044446 <1044446@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1354330 <1354330@85c56323-6fa9-3386-8a01-6480fb634889>