Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Wait a second. I was under the impression for the whole time that this reflects the state of Google3 migration. Is it not?

No and yes. No because it does not jetify (SL -> AndroidX) only to the migrated libraries, it actually does the opposite, it jetifies only to the non migrated, which means that jetifying using the original standalone then dejetifying using this migration.config file means we effectively jetify only the migrated libraries. That is what I'm doing now in g3 once I figure out how to make the config file appear in appear in there.

Patch-set: 2