Add LiveData#switchMap with coroutine context
This CL adds an extension function to LiveData which receives
a CoroutineContext and a block that runs on LiveDataScope.

It is similar to the sequence build where the provided
block can `yield` values that will be dispatched to downstream.

If source LiveData dispatches new values while the block is
running, the existing block's Job is cancelled and a new block
is launched. Even if the previous block does not cooperate
in cancelation, any value it yields will be ignored.

Bug: 122740811
Test: CoroutineLiveDataSwitchMapTest

Change-Id: I727dcfd2386f42c3784d343c579fd58fc9a7af84