Fix Dialogs wiping out Activity Night Mode
Happens when an Activity has a local night mode,
and then a Dialog is created. Since the Dialog has it's
own 'night mode', it will wipe out the Activity mode.

Fixed by just copying the Activity local night in the
Dialog. Then the Dialog's night mode switch will no-op.

Night mode should not really work in Dialogs and we should
remove it, but that's a breaking change which will have
to wait for v2.0.0.

Test: included and passing
BUG: 123566853

Change-Id: I0e8fc52c757a0d135c6b448638e6787296167d4a