Room Suspending Transaction
Create an API for performing database transaction with coroutines. When
starting a transaction coroutine a thread of the database query executor
is hold for performing all database operations within the coroutine. It
is recommended that all DAO functions invoked are also suspend
functions. In the case they are blocking DAO methods, then invoking
those on a different context other than the transaction context will
throw an exception, preventing possible deadlocks.

Bug: 120854786
Test: ./gradlew room:room-compiler:test
                room:integration-tests:room-testapp-kotlin:cC
                room:integration-tests:room-testapp:cC

Change-Id: I30f76db29a3882968f43e1dd1c1038b5afd7f299