Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > From the documentation:
> > Returns the value of this provider if it has a value present, otherwise throws java.lang.IllegalStateException.
> > 
> > There is also getOrElse​(T defaultValue) and getOrNull(), so get() is the best I assume.
> 
> Well -- if this passes then I'm optimistic that we're not accidentally relying on any implicit ordering of resolution. Are you reasonably confident that presubmit is hitting all of the affected code paths?

Not quite certain, I'm assuming it's the same behavior we are getting. This is not the first time we use it though, I've done that in at least one CL in the past where I saw the opportunity to include the swap (.javaCompile -> .javaCompileProvider.get()) and reduce the number of deprecation warnings.

Patch-set: 1
Reviewer: Gerrit User 1354330 <1354330@85c56323-6fa9-3386-8a01-6480fb634889>