Move to the task provider API
This CL moves some of our tasks to the new Provider API, which
avoids configuring the task unless necessary. We have a lot of
tasks that are expensive to configure (doing dependency resolution)
so avoiding this is actually a sizable win for our build.

I've also replaced some usages of `tasks.all` which triggers
realization of all tasks as soon as they are added. We still
have some of these but they are better isolated at this point.

Unfortunately, the CL is a mixed bag of many things as touching
one file requires updating other parts.

Bug: 121135187
Test: https://paste.googleplex.com/5871126655795200
Change-Id: I177c3fc1ffc36afd9e076e379ba4dba000eaf25c