Support parsing inner classes name as types in Safe Args.
This allows using an inner class as arg type if correctly declared using
the dollar sign, e.g. 'com.example.sandbox.OutterClass$InnerClass'.

This CL also updates the non-kotlin nav graph in the PluginTest project
to be the same as the kotlin one, giving us parity coverage for both the
kotlin and java versions of the plugin.

Bug: 123736741
Test: ./gradlew navigation:navigation-safe-args-gradle-plugin:test \
                navigation:navigation-safe-args-generator:test

Change-Id: Ie91ffd0903af7df9b2944f586e3cbbc03ec57bd9