Update patch set 2
Patch Set 2:

It doesn't look like there is a test that makes sure that the RV's RecyclerViewAccessibilityDelegate (as stored in mAccessibilityDelegate) is set on ViewHolder.itemView when the developer does not set their own Accessibility delegate on the item. Given that your change modifies this behavior, it seems like something that would be good to verify.  Could you write that test?

Is there a test that verifies that if a developer adds their own delegate, and then removes it, that RV doesn't automatically add the RecyclerViewAccessibilityDelegate's item delegate.  Does it make sense to verify that? Seems like it would be a good idea, but it can be considered strange to write a test that verifies something doesn't happen.

Patch-set: 2
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>