Use new ActivityRecreator class for recreate() calls
There are a number of platform bugs related to recreation
pre API-28. This class adds workarounds so that calls to
recreate() work as expected. This is especially important
for features such as DayNight.

The way this has been integrated means that as long as
your Activity extends AppCompatActivity, our workarounds
will be used.

BUG: 117842936
Test: tested manually using support-v7-demos
Test: all DayNight tests pass

Change-Id: Id7006fe6c360bc2837776a0b985cccef8a69ab8b