Update patch set 10
Patch Set 10: Presubmit-Ready+1

Removed the Configuration class and replaced directly with Executor. Verified with Sumir that this was the original intention of the comments. Have verified with Sumir on ping that all comments are addressed.

Alan's comments regarding autovalue no longer apply to this CL since it's not used. Also, for the autovalue usage in KeyedAppState, the generated API files look correct (autovalue does not generate any classes that affect the public interface - that is part of the beauty of it). Any minor method naming changes in KeyedAppState required to meet the Android API guidelines should be possible within autovalue as well.

Patch-set: 10
Reviewer: Gerrit User 1388071 <1388071@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1