Update patch set 4
Patch Set 4:

This commit broke the scrolling on Android Oreo for the dropdown. If you have lots of items in a spinner and start scrolling it will keep jumping to the selected item automatically. This makes scrolling (almost) impossible.

In the constructor of AppCompatSpinner, before this commit, mPopupContext would be null on Android Oreo. I'm not sure why this was changed by this commit to a non-null value, but because of this the scrolling broke for me.

Patch-set: 4
CC: Gerrit User 1446205 <1446205@85c56323-6fa9-3386-8a01-6480fb634889>