Fix ImageView drawable state workaround for API 21
ImageView retains its own copy of the tint mode that defaults to null,
and has a one-way relationship to the drawable it owns. Drawables,
meanwhile, generally default to SRC_IN.

If a tint mode is not specified in XML, or set via #setImageTintMode(),
the ImageView will be in a state where it's underlying drawable has at
tint mode of SRC_IN, but calls to #getImageTintMode() will return null.

This CL removes the gating of the drawable state workaround on the
presence of a tint mode on the ImageView, which will allow the drawable
state workaround to function for image views in the default tint mode.

This may create additional overhead from repeated calls to setting the
drawable state on API 21, but the alternative is using reflection to
access ImageView#mHasDrawableTintMode, which tracks the state of an
overriden tint mode.

Test: AppCompatImageViewTest on API 21
Fixes: 117428370
Change-Id: I1b67e82fa6f66717840a92925e79b84f51dd4b2c