Update patch set 5
Patch Set 5:

> Patch Set 2:
> 
> (1 comment)
> 
> I'm curious what the individual library owners think about changing from prebuilt dependencies to tip-of-tree dependencies (there was talk about rewriting some prebuilts if needed, to facilitate keeping dependencies as prebuilts) but I don't inherently have an opinion on that myself.

It's not ideal, but the effect on developers would be minimal.

We discussed the possibility of applying the annotation changes in the release branch and cutting bugfix releases with appropriate restriction scoping, which might be the easiest (and most externally-consistent) approach.

Otherwise, it sounds like rewriting prebuilts and ironing out bugs may take several weeks -- unless you already have a good idea of how we'd write the tooling?

Patch-set: 5