Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Tyson, with this CL I'm seeing an exception trying to resolve dependencies for :room:room-guava:releaseRuntimeClasspath when I update to 27.0.1
> 
> Should I even be updating room to 27, and have you looked at this yet?

I haven't looked at it. The later dep should be OK in room-guava, but there were definitely oddities with the guava and guava-android mirroring.

I've also checked for type mirroring in this CL's classes, and that doesn't seem to be the problem. Got the detailed error? I'm not a Gradle expert, unfortunately.

Patch-set: 1
Reviewer: Gerrit User 1274960 <1274960@85c56323-6fa9-3386-8a01-6480fb634889>