Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What is this a fix for exactly? Can you file a bug about stuff being broken? What is the goal of keeping @Keep? It has retention CLASS right now.

So I can tell you what we see and what my theory is. Going to try to confirm today.

For purely obfuscation reasons, the rules as-is work fine. Obfuscation ignores them.

What we see is that, when enabling optimizations on proguard, the rules aren't respected unless we add the keep for the annotation itself. 

We see this in the form of the Firebase SDK crashing due to the removal of a `@Keep`-annotated method.

	E  java.lang.RuntimeException: Unable to get provider com.google.firebase.provider.FirebaseInitProvider: java.lang.IllegalStateException: com.google.firebase.iid.FirebaseInstanceId#getInstance has been removed by Proguard. Add keep rule to prevent it.
	
My theory for this is that because optimizations run in multiple passes, an earlier pass removes the keep annotation because it's not specified as something that should be kept anywhere. Then a subsequent pass removes the method because the keep annotation is now gone.

Going to try to confirm this by seeing if a single pass optimization doesn't cause it

Patch-set: 1
Reviewer: Gerrit User 1438852 <1438852@85c56323-6fa9-3386-8a01-6480fb634889>