Add tests to the ViewPager2 demo project
Adds a test for every demo activity, and adds several Espresso tools
that may be reused by devs.

Extended and simplified the build.gradle of the demo project.

Moved the demo project to viewpager2/integration-tests/testapp so the
apk of the demo is picked up and installed in our presubmit, so we can
pass our presubmit tests.

Bug: 122659289
Test: ./gradlew viewpager2-demos:connectedCheck
Change-Id: Ia94b813d6b6af620117b65186eb8a928e2cc5980