Deprecate RVA11yDelegate and LayoutManager accessibility functions.
Nearly every client of RecyclerViewAccessibilitYDelegate is doing something we don't want them to do in API (tracking focus, changing row count semantics),
work around bugs that existed at one point or another in frameworks,
or doing something that is much more easily done in newer API (eg addAccessibilityAction).
Getting rid of this needlessly complex API reduces the ways a developer can shoot themselves in the foot, and makes it easier for them to make their recyclerViews accessible.

For example, if the user used the old AXDelegate API, and modified the item delegate, they would need to keep track of layoutManager state to make sure a crash doesn't happen on animation.
With the new API they follow the same pattern as adding an onClickHandler and they avoid that class of bug completely.

Let's move the scroll action code out of the layoutManagers into the RecyclerView proper. Using canScrollHorizontally/Vertically(int) we can deterine what actions should be added. These functions boil down into calls of the layoutManagers.
Now, LayoutManagers don't need to manage scroll accessibility actions themselves (which has been another source of bugs).

While we are here let's get rid of the concept of a the recycler view not generating anything in accessibility because of pending adapter updates. Accessibility reflects the current state of the view hierarchy, not if the hierarchy is attached to a model.
This just seems to add complexity, inconsistency, and potentially some weird bugs.

I added tests to demonstrate adding actions is functional and easy.

Test: Tested with support-v7-demos, autoscroll, two finger scroll, next previous scroll with linear layout, grid, and staggered grid layout managers. RecyclerViewAccessibilityTests (fixing test on this, DO NOT MERGE until this test is fixed).

Change-Id: I13446b02d52653600c02bc6af19294840e281b03