Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> not to self: test launching this inside another multiple times to ensure we don't end up in a cancelled state of a parent is cancelled as parent cancellation will cancel children.

yea so, because this is a scope; there is no way to run something inside another scope. e.g. if you do:

    val job = parentScope.launch {
        withContext(lifecycleScope.coroutineContext) {
           delay(10000)
        }
    }
    delay(200)//wait for with context
    job.cancelAndJoin()

the inner delay will never be cancelled because it is using a context that has a SupervisorJob w/o a parent. idk if taking a coroutineContext from a scope is an OK behavior but it is public and people might do that.

We should try to create an API that won't allow the code above as it seems like an easy mistake to make. Maybe it should be an IDE warning because idk when `withContext` with a `scope.context` can be the desired behavior (even outside lifecycles).

Alternatively, the other CL with dispatcher might obliviate this use case:
https://android-review.googlesource.com/c/platform/frameworks/support/+/905134

Patch-set: 4
Reviewer: Gerrit User 1060337 <1060337@85c56323-6fa9-3386-8a01-6480fb634889>