MediaItem: Makes all Builder methods visible
Previously, a hidden parent class' public methods had been visible in
both current.txt and Javadoc, but it's not true anymore.

This CL removes the previously hidden common builder class
(MediaItem.BuilderBase) and make the MediaItem.Builder the parent class
of all other builder of MediaItem subclasses.

After this CL, when a new method is added to the MediaItem.Builder,
other builders should override it and change return type accordingly.

Bug: 124660696
Test: $ ./gradlew media2:connectedCheck
      $ media2/version-compat-tests/runtest.sh 1
      $ ./gradlew updateApi      # for checking current.txt
      $ ./gradlew distDokkaDocs  # for checking Javadoc
Change-Id: I1027231a65c45ae6cafcf51bcf16dc49c286430e