Update patch set 4
Patch Set 4:

(2 comments)

What about add and replace? Adding a fragment that is already attached to another fragmentManager is the error for sure. 


As a note, I saw in a bug that you don't want to throw in case when Fragment isn't attached to any fragmentManager, but is removed/detached/shown because of fear of the regression. I would argue that these are developer errors as well. Those operations no-op and it is clearly developers error to call beginTransaction().show(f).commitNow() when f isn't added at all to fragmentManager.

Patch-set: 4
Reviewer: Gerrit User 1098853 <1098853@85c56323-6fa9-3386-8a01-6480fb634889>