Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> @Danny is it known failure of androidx.navigation.safe.args.generator.NavGeneratorTest?
> It imports wrong nullability annotation (android vs androidx)

It is known that it uses the the android annotations and not the androidx one, we need to support both. It shouldn't cause a test failure though. I did talk to Jeremy that there is some trickyness to this migration, since Safe Args depends on navigation-common from prebuilts.

Patch-set: 4
Reviewer: Gerrit User 1258207 <1258207@85c56323-6fa9-3386-8a01-6480fb634889>