Don't reset mScrollHappened if scrolls happened
In ScrollEventAdapter, mScrollHappened was reset to false in a few cases
for internal bookkeeping, even though a scroll had actually happened
before. What actually needed to be checked, was if the state handling in
onScrolled was executed, which is more accurately reflected by whether
or not mDispatchSelected is still set to true.

Also add more comments to explain what is going on and why things are
happening. Comments are still not enough, but it's a step in the right
direction.

Test: ./gradlew viewpager2:connectedCheck
Change-Id: I93b7f38a9275fbb58d0ad2b160505addae3ef6dd

fixup! Don't reset mScrollHappened if scrolls happened