Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (2 comments)
> 
> What about add and replace? Adding a fragment that is already attached to another fragmentManager is the error for sure. 
> 
> 
> As a note, I saw in a bug that you don't want to throw in case when Fragment isn't attached to any fragmentManager, but is removed/detached/shown because of fear of the regression. I would argue that these are developer errors as well. Those operations no-op and it is clearly developers error to call beginTransaction().show(f).commitNow() when f isn't added at all to fragmentManager.

I agree, it is a developer error, but add/attach/replace automatically change the FragmentManager a Fragment is associated with. It is similar to removing a Fragment from one FramentManager and adding it to another. If we change that to throw an error, we could break apps.

There are currently tests in FragmentViewTests to ensure that these methods do not throw when not added to a FragmentManager. I was unaware that they no-op though. Would it make more sense to imply an add for show() or hide()?

Patch-set: 4
Reviewer: Gerrit User 1442279 <1442279@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0