Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> (1 comment)
> 
> overall lgtm, just 1 small comment on what we want to do w/ "alwaysBuild" seems like it sohuld be either in both or none.

According to Aurimas, as long as it's in one of the build targets, we avoid the error associated with the bug. This way, the tests only get run on one, though. This way, they don't have large tests run on them.

Patch-set: 7
Reviewer: Gerrit User 1442661 <1442661@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0