[WebView Support Library] Don't get cause for TimeoutException
This fixes a bug in WebkitUtils. Previously, we would call #getCause()
on TimeoutExceptions and InterruptedExceptions, as we do for
ExecutionException.

This makes sense for ExecutionException (it wraps the real Exception, so
we prefer to just throw that Exception to speed up debugging), but
doesn't make sense for the other types. This was a mistake when the
method was originally written; before this patch, timeouts are not
clearly indicated by TimeoutException (which was the original
motivation).

Bug: 125542275
Test: Manual - hack WebViewCompatTest#testVisualStateCallbackCalled so \
Test: it will timeout, and look at the log output (results are pasted \
Test: to the linked bug).
Change-Id: I6e84768f15783ad94e79af802daaecb81f19fafc